refactor: migrate logic to a custom hook
